Create webpage with 4x4 grid
The grid is centered in the middle of the page. The size of the grid is based on the width of the screen and thus, the grid displays properly (on one screen) when the width is the smaller dimension of the page. The squares are blue and with white borders for display purposes.